feat: upload candidateList to firebase -updated importToFirestore.js to allow command-line arguments - <mail/candidate> are the accepted arguments.
